requirements.txt
Django==1.11.6
django-rest-auth==0.9.2
djangorestframework==3.7.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
pytz==2017.2
six==1.11.0
